tests/util: remove unused code
This removes unused code and move inject_sitecustomize to where it is actually
used.
