+++ tests/commands/test_runner.py
@@ -1,7 +1,32 @@
+import os
 import subprocess
+import sys
 
 from ..base import BaseTestCase
-from ..util import inject_sitecustomize
+
+
+def inject_sitecustomize(path):
+    """Creates a new environment, injecting a ``sitecustomize.py`` module in
+    the current PYTHONPATH.
+
+    :param path: package path containing ``sitecustomize.py`` module, starting
+                 from the ddtrace root folder
+    :returns: a cloned environment that includes an altered PYTHONPATH with
+              the given `sitecustomize.py`
+    """
+    from ddtrace import __file__ as root_file
+    root_folder = os.path.dirname(root_file)
+    # Copy the current environment and replace the PYTHONPATH. This is
+    # required otherwise `ddtrace` scripts are not found when `env` kwarg is
+    # passed
+    env = os.environ.copy()
+    sitecustomize = os.path.join(root_folder, '..', path)
+
+    # Add `boostrap` module so that `sitecustomize.py` is at the bottom
+    # of the PYTHONPATH
+    python_path = list(sys.path) + [sitecustomize]
+    env['PYTHONPATH'] = ':'.join(python_path)[1:]
+    return env
 
 
 class DdtraceRunTest(BaseTestCase):
